requirements.txt
py-cord
plexapi
python-dotenv
motor